pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-qa-system"
version = "0.1.0"
description = "A production-ready Retrieval-Augmented Generation system for document Q&A"
authors = [{ name = "RAG Q&A System" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "jinja2>=3.1.4",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "httpx>=0.27.0",
    "scikit-learn>=1.4.0",
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
